eurostat_demo_V1.0/requirements.txt
langchain==0.0.335
langchainplus-sdk==0.0.20
llama-index==0.8.59
streamlit==1.18.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
replicate==0.18.1
#requests==2.31.0
transformers==4.35.0
sentence-transformers==2.2.2
altair==4.1.0
python-dotenv==1.0.0